+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.16.0
+- Adapter API supports `PUT /adapter/entity/:scopeId/:appId/:appVersion/:entityId/:recId
+-
+
 # 0.15.3
 - JSON input only 2 lines in state create form
 - JSON preview in table